requirements.txt
Django==4.1
djangorestframework==3.13.1
gunicorn==20.1.0
openpyxl==3.0.10
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
